+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.1.1 - [Unreleased]
 
+### Added
+- `keepachangelog-open` for just opening the changelog.
+
 ## [0.1.0] - 2025-08-10
 
 First version supporting the core use-case of adding of entries to